Update: modifiy for some REST API and add more details comments

+++ src/main/java/com/example/FirstProject/FirstProject/Controller/AdminController.java
@@ -2,6 +2,8 @@ package com.example.FirstProject.FirstProject.Controller;
 
 
 import com.example.FirstProject.FirstProject.Entity.Admin;
+import com.example.FirstProject.FirstProject.Entity.Course;
+import com.example.FirstProject.FirstProject.Repository.AdminRepository;
 import com.example.FirstProject.FirstProject.Service.Implementation.AdminImp;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,19 +12,29 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
+    /*
+     * REST API for User
+     */
 @RestController
 public class AdminController {
+
+      /*
+         -------- Autowired  AdminImp, AdminRepository for accessing user data
+       */
 
     @Autowired
     private AdminImp adminImp;
 
-    @GetMapping("/welcome/admin")
-    public String HelloAdmin(){
-        return "Welcome to the School";
-    }
+    @Autowired
+    private AdminRepository adminRepository;
 
-//////////////
+
+    /*--------------------
+           - Get a list of all Admin
+           - return list of all Admin
+          */
 
     @GetMapping("/admins")
     @ResponseStatus(HttpStatus.OK)
@@ -30,7 +42,12 @@ public class AdminController {
         return adminImp.getAllAdmins();
     }
 
-    ///////////////
+    /*------------------
+          - add a new  Admin
+          - @RequestBody  Admin to be saved
+          - @Valid to validation
+          - I used try-catch to check there is not give anu error or Exception when added
+            */
 
     @PostMapping("/admins/add")
     public ResponseEntity<String> addAdmin(@RequestBody @Valid Admin admin) {
@@ -44,21 +61,33 @@ public class AdminController {
         }
     }
 
-    //// DELETE
+    /*----------------------
+         - Delete  Admin with handling Exception
+        */
 
     @DeleteMapping("/admins/delete/{id}")
     public String deleteAdmin(@PathVariable int id){
-        adminImp.deleteAdmin(id);
-        return "Admin deleted";
+        Optional<Admin> AdminFound = adminRepository.findById(id);
+        try {
+            if (AdminFound.isPresent()) {
+                adminImp.deleteAdmin(id);
+                return "Admin id deleted";
+            } else {
+                return "Admin with user id " + id + " not found";
+            }
+        } catch (Exception e) {
+            return "Admin not deleted";
+        }
     }
 
-    //// Update
 
-    @PutMapping("/admins/update/{id}")
+     /*----------------------
+      - Update Admin and I did the Exception in sectionImp class
+     */
+
+        @PutMapping("/admins/update/{id}")
     public String updateAdmin(@PathVariable Integer id, @RequestBody Admin admin){
         return adminImp.UpdateAdmin(id,admin);
     }
-
-
 
 }
